pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "draft-hub"
version = "0.1.0"
description = "Collaborative workspace for drafting and voting on amendments."
readme = "README.md"
authors = [
    { name = "Draft Hub" }
]
requires-python = ">=3.12"
dependencies = [
    "PyJWT==2.8.0",
    "pydantic-settings==2.2.1",
    "fastapi==0.112.0",
    "sqlmodel==0.0.22",
    "alembic==1.13.2",
    "uvicorn==0.30.6",
    "redis==5.0.7",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.7,<4",
]

[tool.setuptools]
py-modules = ["app"]

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.ruff]
line-length = 120
lint.select = ["E", "F", "I"]

[tool.flake8]
max-line-length = 120

[tool.mypy]
plugins = [

]
warn_return_any = false
disallow_untyped_defs = true
show_error_codes = true
no_implicit_optional = true
warn_unused_ignores = true
exclude = ["scripts", "docs", "tests"]
ignore_missing_imports = true
